[Refactor] 벡엔드 API 사용으로 기존 불필요한 관광정보 API 제거
