+++ DaOnGil/app/src/main/AndroidManifest.xml
@@ -16,12 +16,11 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".presentation.test.TestActivity"
+            android:exported="false" />
+        <activity
             android:name=".presentation.search.SearchActivity"
             android:exported="true" />
-        <activity
-            android:name=".presentation.TestActivity"
-            android:exported="false" />
-
         <meta-data
             android:name="com.naver.maps.map.CLIENT_ID"
             android:value="s1kt0zadqe" />
