--- DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/dto/remote/response/areabased/Body.kt
@@ -1,11 +1,0 @@
-package kr.tekit.lion.daongil.data.dto.remote.response.areabased
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Body(
-    val items: Items,
-    val numOfRows: Int,
-    val pageNo: Int,
-    val totalCount: Int
-)
