--- DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/dto/remote/response/areabased/Items.kt
@@ -1,8 +1,0 @@
-package kr.tekit.lion.daongil.data.dto.remote.response.areabased
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Items(
-    val item: List<Item>
-)
