+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/network/service/KorWithService.kt
@@ -1,11 +1,8 @@
 package kr.tekit.lion.daongil.data.network.service
 
-import kr.tekit.lion.daongil.data.dto.remote.response.areabased.AreaBasedSearchResponse
 import kr.tekit.lion.daongil.data.dto.remote.response.areacode.AreaCodeResponse
-import kr.tekit.lion.daongil.data.dto.remote.response.barrierfree.BarrierFreeResponse
 import kr.tekit.lion.daongil.data.dto.remote.response.searchkeyword.SearchKeywordResponse
 import retrofit2.http.GET
-import retrofit2.http.Query
 import retrofit2.http.QueryMap
 
 interface KorWithService {
@@ -14,24 +11,9 @@ interface KorWithService {
         @QueryMap params: Map<String, String>
     ): AreaCodeResponse
 
-    @GET("areaBasedList1")
-    suspend fun getAreaBasedResult(
-        @QueryMap params: Map<String, String>
-    ): AreaBasedSearchResponse
-
     @GET("searchKeyword1")
     suspend fun getSearchByKeywordResult(
         @QueryMap params: Map<String, String>
     ): SearchKeywordResponse
 
-    @GET("detailWithTour1")
-    suspend fun getBarrierFreeInfo(
-        @QueryMap params: Map<String, String>
-    ): BarrierFreeResponse
-
-
-    @GET("detailCommon1")
-    suspend fun getSearchByCommonKeyword(
-        @QueryMap params: Map<String, String>
-    )
 }
