--- DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/BarrierFree.kt
@@ -1,33 +1,0 @@
-package kr.tekit.lion.daongil.domain.model
-
-data class BarrierFree (
-    val audioGuide: String,
-    val auditorium: String,
-    val babySpareChair: String,
-    val bigPrint: String,
-    val blindHandicapEtc: String,
-    val braileBlock: String,
-    val brailePromotion: String,
-    val contentId: String,
-    val elevator: String,
-    val exit: String,
-    val guideHuman: String,
-    val guideSystem: String,
-    val handicapEtc: String,
-    val hearingHandicapEtc: String,
-    val hearingRoom: String,
-    val helpDog: String,
-    val infantsFamilyEtc: String,
-    val lactationRoom: String,
-    val parking: String,
-    val promotion: String,
-    val publicTransport: String,
-    val restroom: String,
-    val room: String,
-    val route: String,
-    val signGuide: String,
-    val stroller: String,
-    val ticketOffice: String,
-    val videoGuide: String,
-    val wheelchair: String
-)
