+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/presentation/test/TestActivity.kt
@@ -1,4 +1,4 @@
-package kr.tekit.lion.daongil.presentation
+package kr.tekit.lion.daongil.presentation.test
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -8,5 +8,6 @@ class TestActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_test)
+
     }
 }
