+++ DaOnGil/build.gradle.kts
@@ -9,5 +9,4 @@ buildscript {
 plugins {
     id("com.android.application") version "8.2.1" apply false
     id("org.jetbrains.kotlin.android") version "1.9.22" apply false
-    id("com.google.gms.google-services") version "4.4.1" apply false
 }
